refactory#34: network 패키지 presnetation 밖으로 이동
