+++ app/src/main/java/com/example/yeongkkuel/network/RetrofitClient.kt
@@ -1,10 +1,10 @@
-package com.example.yeongkkuel.presentation.network
+package com.example.yeongkkuel.network
 
 import android.content.Context
 import com.example.yeongkkuel.presentation.auth.ReissueApiService
 import com.example.yeongkkuel.presentation.auth.TokenManager
 import com.example.yeongkkuel.presentation.login.LoginApiService
-import com.example.yeongkkuel.presentation.network.data.AuthInterceptor
+import com.example.yeongkkuel.network.data.AuthInterceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
