+++ app/src/main/java/com/example/yeongkkuel/network/request/expenditure/ExpenditureTargetRequest.kt
@@ -1,4 +1,4 @@
-package com.example.yeongkkuel.presentation.network.request.expenditure
+package com.example.yeongkkuel.network.request.expenditure
 
 import com.google.gson.annotations.SerializedName
