+++ app/src/main/java/com/example/yeongkkuel/presentation/login/LoginFragment.kt
@@ -15,7 +15,7 @@ import com.example.yeongkkuel.R
 import com.example.yeongkkuel.databinding.FragmentLoginBinding
 import com.example.yeongkkuel.presentation.auth.TokenManager
 import com.example.yeongkkuel.presentation.login.response.KakaoLoginResponse
-import com.example.yeongkkuel.presentation.network.RetrofitClient
+import com.example.yeongkkuel.network.RetrofitClient
 import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import com.kakao.sdk.auth.AuthApiClient
