+++ app/src/main/java/com/example/yeongkkuel/presentation/signup/TermsAgreeFragment.kt
@@ -13,8 +13,8 @@ import com.example.yeongkkuel.R
 import com.example.yeongkkuel.databinding.FragmentTermsAgreeBinding
 import com.example.yeongkkuel.presentation.login.request.TermsAgreeRequest
 import com.example.yeongkkuel.presentation.login.response.TermsAgreeResponse
-import com.example.yeongkkuel.presentation.network.RetrofitClient
-import com.example.yeongkkuel.presentation.network.RetrofitClient.loginApiService
+import com.example.yeongkkuel.network.RetrofitClient
+import com.example.yeongkkuel.network.RetrofitClient.loginApiService
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -53,7 +53,7 @@ class TermsAgreeFragment : Fragment() {
 
         // 회원가입 버튼 클릭
         binding.tvSignUp.setOnClickListener {
-            agreeToTerms(viewModel,RetrofitClient)
+            agreeToTerms(viewModel, RetrofitClient)
         }
     }
