+++ app/src/main/java/com/example/yeongkkuel/presentation/stat/weekly/StatWeeklyViewModel.kt
@@ -2,7 +2,7 @@ package com.example.yeongkkuel.presentation.stat.weekly
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.yeongkkuel.presentation.network.RetrofitClient
+import com.example.yeongkkuel.network.RetrofitClient
 import com.example.yeongkkuel.presentation.util.Colors
 import com.example.yeongkkuel.presentation.util.SpendingCategory
 import com.example.yeongkkuel.presentation.util.Week
